feat: pass Request object to beforeRequestHandlers and requestErrorHandlers functions
Pass the `Request` object used when calling `fetch` to `beforeRequestHandlers` and `requestErrorHandlers` functions. This makes it easier to make changes before sending a request (like adding headers) or when trying to deal with request errors.

BREAKING CHANGE: The first positional parameter of `config.beforeRequestHandlers`, the `URL` object used to construct the `Request` object passed to `fetch`, is replaced with that very `Request` object.

BREAKING CHANGE: The second positional parameter of `config.requestErrorHandlers`, the `URL` object used to construct the `Request` object passed to `fetch`, is replaced with that very `Request` object.

BREAKING CHANGE: The third positional parameter of `config.responseErrorHandlers`, the `URL` object used to construct the `Request` object passed to `fetch`, is removed. Use the `request` property on the second positional parameter, the `RetrieveResponse` object, instead.